+++ Cult of the Lamb/CultOfQoL/Patches/Followers/FollowerPatches.cs
@@ -416,6 +416,26 @@ public static class FollowerPatches
 
     private static bool MassLevelUpRunning { get; set; }
 
+    [HarmonyPrefix]
+    [HarmonyPatch(typeof(SimpleSpineAnimator), nameof(SimpleSpineAnimator.Animate), typeof(string), typeof(int), typeof(bool))]
+    public static bool SimpleSpineAnimator_Animate_Prefix(SimpleSpineAnimator __instance)
+    {
+        if (!Plugin.MassLevelUp.Value) return true;
+        if (!MassLevelUpRunning) return true;
+        if (__instance != PlayerFarming.Instance?.simpleSpineAnimator) return true;
+        return false;
+    }
+
+    [HarmonyPrefix]
+    [HarmonyPatch(typeof(SimpleSpineAnimator), nameof(SimpleSpineAnimator.AddAnimate))]
+    public static bool SimpleSpineAnimator_AddAnimate_Prefix(SimpleSpineAnimator __instance)
+    {
+        if (!Plugin.MassLevelUp.Value) return true;
+        if (!MassLevelUpRunning) return true;
+        if (__instance != PlayerFarming.Instance?.simpleSpineAnimator) return true;
+        return false;
+    }
+
     [HarmonyPostfix]
     [HarmonyPatch(typeof(interaction_FollowerInteraction), nameof(interaction_FollowerInteraction.LevelUpRoutine))]
     public static void MassLevelUp_Postfix(ref interaction_FollowerInteraction __instance)